공통 레트로핏 병합
